% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despike_30min_Papale06.R
\name{despike_30min_Papale06}
\alias{despike_30min_Papale06}
\title{Title}
\usage{
despike_30min_Papale06(flux.df, z = 4)
}
\arguments{
\item{flux.df}{}

\item{z}{}
}
\value{

}
\description{
Title
}
