% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEON_2_REddyProc.R
\name{extract_NEON_fluxes_REddyProc}
\alias{extract_NEON_fluxes_REddyProc}
\title{Title}
\usage{
extract_NEON_fluxes_REddyProc(neon.site, year = 9999,
  flux.path = "~/Dropbox/NEON/DP4_00200_001", met.path,
  expanded = FALSE, median.filter = TRUE, write.to.file = FALSE,
  nee.max, nee.min, lh.max, lh.min)
}
\arguments{
\item{neon.site}{}

\item{year}{}

\item{flux.path}{}

\item{met.path}{}

\item{expanded}{}

\item{median.filter}{}

\item{write.to.file}{}

\item{nee.max}{}

\item{nee.min}{}

\item{lh.max}{}

\item{lh.min}{}
}
\description{
Title
}
