% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met_terms.R
\name{plot_met_terms}
\alias{plot_met_terms}
\title{plot_met_terms}
\usage{
plot_met_terms(data)
}
\arguments{
\item{data}{\code{data.frame} containing the meteorological variables that should be plotted. Requires
\code{data.frame} to be the expanded version from extract_NEON_fluxes.}
}
\value{
Multipanel plot showing Tair, RH, VPD, PAR, and Ustar.
}
\description{
Plots the meteological variable terms extracted from extract_NEON_fluxes
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
