% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nee_vs_par.R
\name{plot_nee_vs_par}
\alias{plot_nee_vs_par}
\title{Title}
\usage{
plot_nee_vs_par(data, nee.thres = 50)
}
\arguments{
\item{data}{Input dataframe. Must have all 9 terms (3 CO2, 3 water, 3 temperature) from the eddy covariance data product in the timeseries.}
}
\value{
Multipanel plot showing all 9 terms (3 net exchange terms, 3 turbulent flux terms, 3 storage terms.)
}
\description{
Title
}
