% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storage_vs_ustar.R
\name{plot_storage_vs_ustar}
\alias{plot_storage_vs_ustar}
\title{Title}
\usage{
plot_storage_vs_ustar(data)
}
\arguments{
\item{data}{Input dataframe. Must have all 9 terms (3 CO2, 3 water, 3 temperature) from the eddy covariance data product in the timeseries.}
}
\value{
Multipanel plot showing all 9 terms (3 net exchange terms, 3 turbulent flux terms, 3 storage terms.)
}
\description{
Title
}
