% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NEON_fluxes.R
\name{extract_NEON_fluxes}
\alias{extract_NEON_fluxes}
\title{Title}
\usage{
extract_NEON_fluxes(
  neon.site,
  year = 9999,
  flux.path = "~/Dropbox/NEON/DP4_00200_001",
  met.path,
  expanded = FALSE,
  median.filter = TRUE,
  write.to.file = FALSE,
  out.path,
  nee.max,
  nee.min,
  lh.max,
  lh.min
)
}
\arguments{
\item{neon.site}{}

\item{year}{}

\item{flux.path}{}

\item{met.path}{}

\item{expanded}{}

\item{median.filter}{}

\item{write.to.file}{}

\item{nee.max}{}

\item{nee.min}{}

\item{lh.max}{}

\item{lh.min}{}
}
\value{

}
\description{
Title
}
