% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NEON_fluxes.R
\name{extract_NEON_fluxes}
\alias{extract_NEON_fluxes}
\title{extract_NEON_fluxes}
\usage{
extract_NEON_fluxes(
  neon.site,
  year = 9999,
  flux.path = "~/Dropbox/NEON/DP4_00200_001",
  met.path,
  expanded = FALSE,
  median.filter = TRUE,
  filt.width = 3,
  fix.tz = FALSE,
  write.to.file = FALSE,
  out.path
)
}
\arguments{
\item{neon.site}{Four letter code indicating NEON site.}

\item{year}{Which year to process? If not specified, process all.}

\item{flux.path}{Specify path to flux data you wish to process.}

\item{met.path}{Specify path to meteorological data (if NULL - download from NEON API)}

\item{expanded}{}

\item{median.filter}{Filter half-hourly data using the Brock 86 median filter.}

\item{filt.width}{Width of median filter. Old implementation.}

\item{fix.tz}{Convert from UTC to local time zone?}

\item{write.to.file}{Write to csv file?}

\item{out.path}{If saving a file of results, where should it be saved?}
}
\value{

}
\description{
extract_NEON_fluxes
}
